Update 'read keystroke' recipe in local-io.
Responded to Ryan's and Colin's comments.  Note that this version has a footnote which, when HTML output is produced, renders at the end of the entire HTML document (i.e. the end of the book), but which should render properly at page bottom in PDF output (which I haven't gotten working yet in my asciidoc installation).