Rationalize all uses of the links to clj-time
Other recipes reference https://github.com/clj-time/clj-time
as the url for the clj-time project, and this source has
been updated 4 days ago as of this commit.  These two files used
https://github.com/KirinDave/clj-time, which has not been updated
in about 3 years, except for the license file which was updated
around a year ago according to the project page.