ClojureScript will, unfortunately, not make it into the first edition.
To those who contributed ClojureScript recipes, I apologize. The
schedule has just been too tight for us to do a good job of the
ClojureScript recipes. I'm canning these recipes from master to reduce
confusion of technical editors and repository visitors.

Rest assured, ClojureScript will return once we start building the book
out as on online resource. I've captured the reversion of this commit in
the branch 'cljs-is-here-to-stay'.