Update 1-05_integer-to-character-conversions.asciidoc
Using reduce with an anonymous function to build up a string from ascii codes is a bit of overkill, and non-intuitive for new users in the first chapter.

I think it is much more straightforward to use a simple "map" call to convert from integer to char, the "apply str" to convert the individual chars into a single string.

I also put the vector of ascii codes into a separate vector "codes" with the idea that 2 simple steps are faster & easier to understand (esp. in Ch #1) than a single large step.