Update 2-13_testing-set-membership.asciidoc
Luke seems to have forgotten that sets are essentially maps where the same item is both the map key and the map value.  In recipe 2.16 he points out the ability to specify a default for both get access and keyword access to the map, but forgot to mention it here for sets.