4-17 Unknown Tagged Literals: Add a See Also item
It took me a long time to figure out why clojure.edn/read didn't heed
data_readers.clj. (Because it's not supposed to.) Looking for answers,
recipe 4-17 turned up quite high in the search results, but didn't help
me much. When I finally figured it out, I hit my head against the wall
and suspected that other people have the same problems. So I added a
note to the ClojureDocs entry for clojure.edn/read.

Add a link to the ClojureDocs entry to the See Also section of recipe
4-17, so that people finding this recipe first, can achieve
enlightenment more quickly.